% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbot_download.R
\name{fishbot_download}
\alias{fishbot_download}
\title{fishbot_download}
\usage{
fishbot_download(EPU = NA, startDate, endDate)
}
\arguments{
\item{EPU}{the EPU of the dataset to download. Acceptable values include
'GOM' for Gulf of Maine, 'GB' for Georges Bank, or 'MAB' for Mid-Atlantic
Bight. The default value is NA (no bounding box applied).}

\item{startDate}{The earliest date to return. Must be provided in the format yyyy-mm-dd.
No default value is provided.}

\item{endDate}{The latest date to return. Must be provided in the format yyyy-mm-dd.
No default value is provided.}
}
\value{
Returns a dataframe including the following variables
\itemize{
\item time (yyyy-mm-dd)
\item temperature (average bottom temperature of a cell in degrees C)
\item temperature_min (minimum observed bottom temperature of a cell in degrees C)
\item tempertaure_max (maximum observed bottom temperature of a cell in degrees C)
\item temperature_std (standard deviation of bottom temperature in a cell in degrees C)
\item temperature_count (number of observations of bottom temperature in a cell)
\item data_source (which program contributed the observations)
\item stat_area (NOAA statistical area the cell is in)
\item depth (GEBCO depth of the cell in meters)
\item latitude (NAD83 decimal degrees)
\item longitude (NAD83 decimal degrees)
}
}
\description{
First pass at this downloads a .csv file within a date range and bounding box
}
\examples{

fishbot_download(
  EPU = 'GOM',
  startDate = '2023-08-01',
  endDate = '2023-08-31'
  )
  
}
