% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIShBOTR-package.R
\docType{package}
\name{FIShBOTR-package}
\alias{FIShBOTR}
\alias{FIShBOTR-package}
\title{FIShBOTR: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GMaynard1/FIShBOTR}
  \item Report bugs at \url{https://github.com/GMaynard1/FIShBOTR/issues}
}

}
\author{
\strong{Maintainer}: George Maynard \email{george.maynard@noaa.gov} (\href{https://orcid.org/0000-0003-4246-2345}{ORCID}) [copyright holder]

}
\keyword{internal}
