% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_select.R
\name{grid_select}
\alias{grid_select}
\title{grid_select}
\usage{
grid_select(data)
}
\arguments{
\item{data}{a dataframe that includes at least the parameters latitude and
longitude in decimal degrees}
}
\value{
Returns the dataframe with an additional column of index values
}
\description{
Uses latitude and longitude points to select grid id values
}
\examples{

  dat=data.frame(
    site_names = c("Skate Bank","Cholera Bank"),
    latitude = c(43.611,40.395),
    longitude = c(-68.646,-73.581)
    )
  grid_select(df)
  
  @export
}
