% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_trim.R
\name{grid_trim}
\alias{grid_trim}
\title{grid_trim}
\usage{
grid_trim(clip_shape)
}
\arguments{
\item{clip_shape}{can either be a polygon shapefile (NAD83), a bounding box in the format
c(min_lat, max_lat, min_lon, max_lon) with all coordinates expressed in decimal
degrees, a character vector of statistical areas in the format
c("SA-514","SA-521") or a character vector of a single Ecosystem Production Unit
(EPU) that can be one of the following: "EPU-GOM", "EPU-GB",
"EPU-SS" or "EPU-MAB" for Gulf of Maine, Georges Bank, Scotian Shelf, and
Mid-Atlantic Bight respectively. Note that the clip is
exact. Grid cells that are intersected by the clip_shape will be
proportioned in subsequent analyses. There is no default value for clip_shape.}
}
\value{
Returns a trimmed version of the FIShBOT grid as a spatial dataframe
}
\description{
Uses a shapefile or bounding box to trim the FIShBOT grid to the specified size
}
\examples{

grid_trim(
  clip_shape=c(42,43,-72,-71)
  )
  
}
